Update diagrams from sanitizer service
So they will also be updated if there's no diagrams component opened.
